extract legacy reconstruction module
